requirements.txt
Flask==1.1.2
numpy
tensorflow
matplotlib
opencv-python
Pillow
keras==2.4.3